Refactor: extract insertSectionAt logic into a separate utility function
